move nsstring extension to new file
