+++ wire-ios-utilities/Source/Public/String+Random.swift
@@ -18,34 +18,7 @@
 
 import Foundation
 
-extension NSString {
-
-    @objc static func randomAlphanumerical(length: UInt) -> String {
-        String.randomAlphanumerical(length: length)
-    }
-
-    @objc public static func randomClientIdentifier() -> String {
-        String.randomClientIdentifier()
-    }
-
-    @objc public static func randomRemoteIdentifier() -> String {
-        String.randomRemoteIdentifier()
-    }
-}
-
 public extension String {
-
-    static func randomClientIdentifier(length: UInt = 16) -> String {
-        randomAlphanumerical(length: length)
-    }
-
-    static func randomRemoteIdentifier(length: UInt = 16) -> String {
-        randomAlphanumerical(length: length)
-    }
-
-    static func randomDomain(hostLength: UInt = 5) -> String {
-        return "\(String.randomAlphanumerical(length: hostLength)).com"
-    }
 
     static func randomAlphanumerical(length: UInt) -> String {
         if length == 0 {
@@ -60,4 +33,15 @@ public extension String {
         return s
     }
 
+    static func randomClientIdentifier(length: UInt = 16) -> String {
+        randomAlphanumerical(length: length)
+    }
+
+    static func randomDomain(hostLength: UInt = 5) -> String {
+        return "\(String.randomAlphanumerical(length: hostLength)).com"
+    }
+
+    static func randomRemoteIdentifier(length: UInt = 16) -> String {
+        randomAlphanumerical(length: length)
+    }
 }
